% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_venn_NAs.R
\name{geom_venn_NAs}
\alias{geom_venn_NAs}
\title{Create a Venn Diagram for Missing Values}
\usage{
geom_venn_NAs(var1, var2, fill = c("#FF6F91", "#CC4C6B"), ...)
}
\arguments{
\item{var1}{A vector or data frame column to be evaluated for missing values.}

\item{var2}{A vector or data frame column to be evaluated for missing values.}

\item{fill}{A character vector of two colors used to fill the individual and intersection sets.
Defaults to `c("#FF6F91", "#CC4C6B")`.}
}
\value{
A ggplot object with an overlaid Venn diagram displaying the count of missing values in
`var1`, `var2`, and their overlap.
}
\description{
`geom_venn_NAs` generates a Venn diagram highlighting the overlap of missing values
between two input variables.
}
\examples{
# Example usage with dummy data
var1 <- c(1, 2, NA, 4, NA)
var2 <- c(NA, 2, NA, 4, 5)
geom_venn_NAs(var1, var2)

}
