% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairsNAs.R
\name{ggpairsNAs}
\alias{ggpairsNAs}
\title{Pairwise Plot of Missing Values}
\usage{
ggpairsNAs(data, fill = c("#01BFC4", "#007a7c"))
}
\arguments{
\item{data}{A data frame containing variables to be analyzed for missing values.}

\item{fill}{A character vector of two colors used to fill the visual elements in the Venn and bar diagrams.
Defaults to `c("#01BFC4", "#007a7c")`.}
}
\value{
A `ggpairs` object displaying pairwise plots of missing values for each variable in `data`.
}
\description{
`ggpairsNAs` creates a matrix of pairwise plots for visualizing missing values across variables in a data frame.
The function uses customized upper, lower, and diagonal panels to display missing value counts and relationships.
}
\details{
The function customizes the following:
- **Diagonal Panel**: Displays the count of missing values for each variable.
- **Lower Panel**: Displays a Venn diagram showing the overlap of missing values between variable pairs.
- **Upper Panel**: Shows a bar plot with counts of missing values for each variable.
}
\examples{
# Example usage with dummy data
data <- data.frame(
  var1 = c(1, 2, NA, 4, NA),
  var2 = c(NA, 2, NA, 4, 5),
  var3 = c(1, NA, 3, 4, 5)
)
ggpairsNAs(data)

}
